pypi: use local persistent volume for storage
This is a temporary measure while external persistent volumes are not configured
in the Kubernetes cluster.  While it ties the pod to a specific node, it removes
the need to recreate the package archive each time the pod is recreated or
restarted.

The package directory is changed because the `Dockerfile` contains a `VOLUME`
instruction for `/data/packages`:

https://github.com/pypiserver/pypiserver/blob/01915d1a689327da763efb3a3358f4e91207da24/Dockerfile#L77